Convert size_t into unsigned int in test files.

+++ tests/errors_ut.cpp
@@ -21,7 +21,7 @@ class ErrorMessageListener {
         ~ErrorMessageListener();
 
         std::string getCurrentErrorMsg();
-        std::string getLine(size_t);
+        std::string getLine(unsigned int);
 };
 
 /**
@@ -61,7 +61,7 @@ std::string ErrorMessageListener::getCurrentErrorMsg() {
 /**
 * Returns a line from the currently read stderr message.
 */
-std::string ErrorMessageListener::getLine(size_t lineNumber) {
+std::string ErrorMessageListener::getLine(unsigned int lineNumber) {
     auto errorMessage = std::istringstream(this->getCurrentErrorMsg());
     std::string errorLine = "";
     for (auto i = 0; i < lineNumber; i++) {
@@ -97,28 +97,28 @@ int main(int argc, const char** argv) {
                           stderrListener = std::make_unique<ErrorMessageListener>(std::string(syntaxException.what()));
                       }
 
-                      steps.then("line {lineNumber} of the error message should mention '{errorFound}'.") = [&](size_t lineNumber, std::string errorFound) {
+                      steps.then("line {lineNumber} of the error message should mention '{errorFound}'.") = [&](unsigned int lineNumber, std::string errorFound) {
                           auto expectedErrorSummary = std::format("Error: {}", errorFound);
                           auto actualErrorSummary = stderrListener->getLine(lineNumber);
 
                           expect_eq(expectedErrorSummary, actualErrorSummary);
                       };
 
-                      steps.then("line {lineNumber} of the error message should mention finding {errorFound} in the wrong spot.") = [&](size_t lineNumber, std::string errorFound) {
+                      steps.then("line {lineNumber} of the error message should mention finding {errorFound} in the wrong spot.") = [&](unsigned int lineNumber, std::string errorFound) {
                           auto expectedErrorSummary = std::format("Error: A {} is not allowed in its current spot.", errorFound);
                           auto actualErrorSummary = stderrListener->getLine(lineNumber);
 
                           expect_eq(expectedErrorSummary, actualErrorSummary);
                       };
 
-                      steps.then("line {lineNumber} of the error message should mention the given TSL input file at line {inputLineNumber}, column {inputColumnNumber}.") = [&](size_t lineNumber, size_t inputLineNumber, size_t inputLineColumn) {
+                      steps.then("line {lineNumber} of the error message should mention the given TSL input file at line {inputLineNumber}, column {inputColumnNumber}.") = [&](unsigned int lineNumber, unsigned int inputLineNumber, unsigned int inputLineColumn) {
                           auto expectedFileLine = std::format(" --> {}:{}.{}", parser.getLexer().getFileName(), inputLineNumber, inputLineColumn);
                           auto actualFileLine = stderrListener->getLine(lineNumber);
 
                           expect_eq(expectedFileLine, actualFileLine);
                       };
 
-                      steps.then("line {lineNumber} of the error message should point to the token {errorToken} in line {expectedFirstLineNumber} with line {expectedNextLineNumber} below it.") = [&](size_t lineNumber, std::string errorToken, size_t expectedFirstLineNumber, size_t expectedNextLineNumber) {
+                      steps.then("line {lineNumber} of the error message should point to the token {errorToken} in line {expectedFirstLineNumber} with line {expectedNextLineNumber} below it.") = [&](unsigned int lineNumber, std::string errorToken, unsigned int expectedFirstLineNumber, unsigned int expectedNextLineNumber) {
                           auto actualErrorPointedOut = stderrListener->getLine(lineNumber) + "\n"
                             + stderrListener->getLine(lineNumber + 1) + "\n";
